+++ bertopic/__init__.py
@@ -2,7 +2,7 @@ from bertopic._bertopic import BERTopic
 from bertopic._ctfidf import ClassTFIDF
 from bertopic._embeddings import languages
 
-__version__ = "0.4.3"
+__version__ = "0.5.0"
 
 __all__ = [
     "BERTopic",