+++ docs/changelog.md
@@ -1,3 +1,21 @@
+## **Version 0.5.0**
+*Release date:  8 Februari, 2021*
+
+**Highlights**:
+  
+* Add `Flair` to allow for more (custom) token/document embeddings, including 🤗 transformers 
+* Option to use custom UMAP, HDBSCAN, and CountVectorizer
+* Added `low_memory` parameter to reduce memory during computation
+* Improved verbosity (shows progress bar)
+* Return the figure of `visualize_topics()`
+* Expose all parameters with a single function: `get_params()`
+
+**Fixes**:
+    
+* To simplify the API, the parameters stop_words and n_neighbors were removed. These can still be used when a custom UMAP or CountVectorizer is used.
+* Set `calculate_probabilities` to False as a default. Calculating probabilities with HDBSCAN significantly increases computation time and memory usage. Better to remove calculating probabilities or only allow it by manually turning this on.
+* Use the newest version of `sentence-transformers` as it speeds ups encoding significantly
+
 ## **Version 0.4.2**
 *Release date:  10 Januari, 2021*